refactor(pr): Move getResourceSummariesFor() back to ApplicationService
